+++ tests/phpunit/ArrayComparer/StrategicArrayComparerTest.php
@@ -2,6 +2,7 @@
 
 namespace Diff\Tests\ArrayComparer;
 
+use Diff\ArrayComparer\ArrayComparer;
 use Diff\ArrayComparer\StrategicArrayComparer;
 use Diff\Tests\DiffTestCase;
 
@@ -9,7 +10,7 @@ use Diff\Tests\DiffTestCase;
  * @covers Diff\ArrayComparer\StrategicArrayComparer
  *
  * @file
- * @since 0.7
+ * @since 0.8
  *
  * @ingroup DiffTest
  *
@@ -25,26 +26,179 @@ class StrategicArrayComparerTest extends DiffTestCase {
 		$this->assertTrue( true );
 	}
 
-	public function testDiffArrays() {
+	public function testDiffArraysWithComparerThatAlwaysReturnsTrue() {
+		$valueComparer = $this->getMock( 'Diff\Comparer\ValueComparer' );
+
+		$valueComparer->expects( $this->any() )
+			->method( 'valuesAreEqual' )
+			->will( $this->returnValue( true ) );
+
+		$arrayComparer = new StrategicArrayComparer( $valueComparer );
+
+		$this->assertNoDifference(
+			$arrayComparer,
+			array( 0, 2, 4 ),
+			array( 1, 2, 9 )
+		);
+
+		$this->assertNoDifference(
+			$arrayComparer,
+			array( 1, 2, 3 ),
+			array( 1, 2, 3 )
+		);
+
+		$this->assertNoDifference(
+			$arrayComparer,
+			array( 'bah' ),
+			array( 'foo', 'bar', 'baz' )
+		);
+
+		$this->assertNoDifference(
+			$arrayComparer,
+			array(),
+			array( 'foo', 'bar', 'baz' )
+		);
+
+		$this->assertNoDifference(
+			$arrayComparer,
+			array(),
+			array()
+		);
+	}
+
+	protected function assertNoDifference( ArrayComparer $arrayComparer, array $arrayOne, array $arrayTwo ) {
+		$this->assertEquals(
+			array(),
+			$arrayComparer->diffArrays(
+				$arrayOne,
+				$arrayTwo
+			)
+		);
+	}
+
+	public function testDiffArraysWithComparerThatAlwaysReturnsFalse() {
+		$valueComparer = $this->getMock( 'Diff\Comparer\ValueComparer' );
+
+		$valueComparer->expects( $this->any() )
+			->method( 'valuesAreEqual' )
+			->will( $this->returnValue( false ) );
+
+		$arrayComparer = new StrategicArrayComparer( $valueComparer );
+
+		$this->assertAllDifferent(
+			$arrayComparer,
+			array(),
+			array()
+		);
+
+		$this->assertAllDifferent(
+			$arrayComparer,
+			array( 1, 2, 3 ),
+			array()
+		);
+
+		$this->assertAllDifferent(
+			$arrayComparer,
+			array( 1, 2, 3 ),
+			array( 1, 2, 3 )
+		);
+
+		$this->assertAllDifferent(
+			$arrayComparer,
+			array(),
+			array( 1, 2, 3 )
+		);
+	}
+
+	protected function assertAllDifferent( ArrayComparer $arrayComparer, array $arrayOne, array $arrayTwo ) {
+		$this->assertEquals(
+			$arrayOne,
+			$arrayComparer->diffArrays(
+				$arrayOne,
+				$arrayTwo
+			)
+		);
+	}
+
+	public function testQuantityMattersWithReturnTrue() {
+		$valueComparer = $this->getMock( 'Diff\Comparer\ValueComparer' );
+
+		$valueComparer->expects( $this->any() )
+			->method( 'valuesAreEqual' )
+			->will( $this->returnValue( true ) );
+
+		$arrayComparer = new StrategicArrayComparer( $valueComparer );
+
+		$this->assertEquals(
+			array( 1, 1, 1 ),
+			$arrayComparer->diffArrays(
+				array( 1, 1, 1, 1 ),
+				array( 1 )
+			)
+		);
+
+		$this->assertEquals(
+			array( 1 ),
+			$arrayComparer->diffArrays(
+				array( 1, 1, 1, 1 ),
+				array( 1, 1, 1  )
+			)
+		);
+	}
+
+	public function testQuantityMattersWithSimpleComparison() {
 		$valueComparer = $this->getMock( 'Diff\Comparer\ValueComparer' );
 
 		$valueComparer->expects( $this->any() )
 			->method( 'valuesAreEqual' )
 			->will( $this->returnCallback( function( $firstValue, $secondValue ) {
-				return true;
+				return $firstValue == $secondValue;
 			} ) );
 
 		$arrayComparer = new StrategicArrayComparer( $valueComparer );
 
 		$this->assertEquals(
-			array(),
+			array( 1, 2, 5 ),
 			$arrayComparer->diffArrays(
-				array( 0, 2, 4 ),
-				array( 1, 2, 9 )
+				array( 1, 1, 2, 3, 2, 5 ),
+				array( 1, 2, 3, 4  )
 			)
 		);
 
-		// TODO: implement
+		$this->assertEquals(
+			array( 1 ),
+			$arrayComparer->diffArrays(
+				array( 1, 1, 1, 2, 2, 3 ),
+				array( 1, 1, 2, 2, 3, 3, 3 )
+			)
+		);
+
+		$this->assertEquals(
+			array( 1 ),
+			$arrayComparer->diffArrays(
+				array( 3, 1, 2, 1, 1, 2 ),
+				array( 1, 3, 3, 2, 2, 3, 1 )
+			)
+		);
+	}
+
+	public function testValueComparerGetsCalledWithCorrectValues() {
+		$valueComparer = $this->getMock( 'Diff\Comparer\ValueComparer' );
+
+		$valueComparer->expects( $this->once() )
+			->method( 'valuesAreEqual' )
+			->with(
+				$this->equalTo( 1 ),
+				$this->equalTo( 2 )
+			)
+			->will( $this->returnValue( true ) );
+
+		$arrayComparer = new StrategicArrayComparer( $valueComparer );
+
+		$arrayComparer->diffArrays(
+			array( 1 ),
+			array( 2 )
+		);
 	}
 
 }