+++ tests/phpunit/ArrayComparer/StrictArrayComparerTest.php
@@ -9,7 +9,7 @@ use Diff\Tests\DiffTestCase;
  * @covers Diff\ArrayComparer\StrictArrayComparer
  *
  * @file
- * @since 0.7
+ * @since 0.8
  *
  * @ingroup DiffTest
  *